requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
python-dotenv
ollama
streamlit
PyPDF2